+++ bitstring/__init__.py
@@ -177,27 +177,44 @@ def bool_bits2chars(_: Literal[1]):
     # Bools are printed as 1 or 0, not True or False, so are one character each
     return 1
 
+byteorder: str = sys.byteorder
+
+native_setuint = Bits._setuintle if byteorder == 'little' else Bits._setuintbe
+native_getuint = Bits._getuintle if byteorder == 'little' else Bits._getuintbe
+native_readuint = Bits._readuintle if byteorder == 'little' else Bits._readuintbe
+native_setint = Bits._setintle if byteorder == 'little' else Bits._setintbe
+native_getint = Bits._getintle if byteorder == 'little' else Bits._getintbe
+native_readint = Bits._readintle if byteorder == 'little' else Bits._readintbe
+native_setfloat = Bits._setfloatle if byteorder == 'little' else Bits._setfloatbe
+native_getfloat = Bits._getfloatle if byteorder == 'little' else Bits._getfloatbe
+native_readfloat = Bits._readfloatle if byteorder == 'little' else Bits._readfloatbe
 
 dtype_definitions = [
     # Integer types
     DtypeDefinition('u', Bits._setuint, Bits._getuint, int, False, uint_bits2chars,
                     read_fn=Bits._readuint,
                     description="a two's complement unsigned int"),
-    DtypeDefinition('uintle', Bits._setuintle, Bits._getuintle, int, False, uint_bits2chars,
+    DtypeDefinition('ule', Bits._setuintle, Bits._getuintle, int, False, uint_bits2chars,
                     read_fn=Bits._readuintle,
                     allowed_lengths=(8, 16, 24, ...), description="a two's complement little-endian unsigned int"),
-    DtypeDefinition('uintbe', Bits._setuintbe, Bits._getuintbe, int, False, uint_bits2chars,
+    DtypeDefinition('ube', Bits._setuintbe, Bits._getuintbe, int, False, uint_bits2chars,
                     read_fn=Bits._readuintbe,
                     allowed_lengths=(8, 16, 24, ...), description="a two's complement big-endian unsigned int"),
+    DtypeDefinition('une', native_setuint, native_getuint, int, False, uint_bits2chars,
+                    read_fn=native_readuint,
+                    allowed_lengths=(8, 16, 24, ...), description="a two's complement native-endian unsigned int"),
     DtypeDefinition('i', Bits._setint, Bits._getint, int, True, int_bits2chars,
                     read_fn=Bits._readint,
                     description="a two's complement signed int"),
-    DtypeDefinition('intle', Bits._setintle, Bits._getintle, int, True, int_bits2chars,
+    DtypeDefinition('ile', Bits._setintle, Bits._getintle, int, True, int_bits2chars,
                     read_fn=Bits._readintle,
                     allowed_lengths=(8, 16, 24, ...), description="a two's complement little-endian signed int"),
-    DtypeDefinition('intbe', Bits._setintbe, Bits._getintbe, int, True, int_bits2chars,
+    DtypeDefinition('ibe', Bits._setintbe, Bits._getintbe, int, True, int_bits2chars,
                     read_fn=Bits._readintbe,
                     allowed_lengths=(8, 16, 24, ...), description="a two's complement big-endian signed int"),
+    DtypeDefinition('ine', native_setint, native_getint, int, True, int_bits2chars,
+                    read_fn=native_readint,
+                    allowed_lengths=(8, 16, 24, ...), description="a two's complement native-endian signed int"),
     # String types
     DtypeDefinition('hex', Bits._sethex, Bits._gethex, str, False, hex_bits2chars,
                     read_fn=Bits._readhex,
@@ -212,9 +229,12 @@ dtype_definitions = [
     DtypeDefinition('f', Bits._setfloatbe, Bits._getfloatbe, float, True, float_bits2chars,
                     read_fn=Bits._readfloatbe,
                     allowed_lengths=(16, 32, 64), description="a big-endian floating point number"),
-    DtypeDefinition('floatle', Bits._setfloatle, Bits._getfloatle, float, True, float_bits2chars,
+    DtypeDefinition('fle', Bits._setfloatle, Bits._getfloatle, float, True, float_bits2chars,
                     read_fn=Bits._readfloatle,
                     allowed_lengths=(16, 32, 64), description="a little-endian floating point number"),
+    DtypeDefinition('fne', native_setfloat, native_getfloat, float, True, float_bits2chars,
+                    read_fn=native_readfloat,
+                    allowed_lengths=(16, 32, 64), description="a native-endian floating point number"),
     DtypeDefinition('bfloat', Bits._setbfloatbe, Bits._getbfloatbe, float, True, bfloat_bits2chars,
                     read_fn=Bits._readbfloatbe,
                     allowed_lengths=(16,), description="a 16 bit big-endian bfloat floating point number"),
@@ -279,27 +299,31 @@ aliases: list[tuple[str, str]] = [
     # Floats default to big endian
     ('f', 'float'),
     ('f', 'floatbe'),
+    ('f', 'fbe'),
     ('bfloat', 'bfloatbe'),
 
     # Longer compatibility aliases for the short core type names
     ('i', 'int'),
     ('u', 'uint'),
+
+    # Longer compatibility aliases for endian-specific type names
+    ('ube', 'uintbe'),
+    ('ule', 'uintle'),
+    ('une', 'uintne'),
+    ('ibe', 'intbe'),
+    ('ile', 'intle'),
+    ('ine', 'intne'),
+    ('fle', 'floatle'),
+    ('fne', 'floatne'),
 ]
 
 # Create native-endian aliases depending on the byteorder of the system
-byteorder: str = sys.byteorder
 if byteorder == 'little':
     aliases.extend([
-        ('uintle', 'uintne'),
-        ('intle', 'intne'),
-        ('floatle', 'floatne'),
         ('bfloatle', 'bfloatne'),
     ])
 else:
     aliases.extend([
-        ('uintbe', 'uintne'),
-        ('intbe', 'intne'),
-        ('floatbe', 'floatne'),
         ('bfloatbe', 'bfloatne'),
     ])