+++ bitstring/bits.py
@@ -84,23 +84,23 @@ class Bits:
         bytes -- raw data as a bytes object, for example read from a binary file.
         i -- a signed integer. int is a compatibility alias.
         u -- an unsigned integer. uint is a compatibility alias.
-        f / float / floatbe -- a big-endian floating point number.
+        f / float / fbe -- a big-endian floating point number.
         bool -- a boolean (True or False).
         se -- a signed exponential-Golomb code.
         ue -- an unsigned exponential-Golomb code.
         sie -- a signed interleaved exponential-Golomb code.
         uie -- an unsigned interleaved exponential-Golomb code.
-        floatle -- a little-endian floating point number.
-        floatne -- a native-endian floating point number.
+        fle -- a little-endian floating point number.
+        fne -- a native-endian floating point number.
         bfloat / bfloatbe - a big-endian bfloat format 16-bit floating point number.
         bfloatle -- a little-endian bfloat format 16-bit floating point number.
         bfloatne -- a native-endian bfloat format 16-bit floating point number.
-        intbe -- a signed big-endian whole byte integer.
-        intle -- a signed little-endian whole byte integer.
-        intne -- a signed native-endian whole byte integer.
-        uintbe -- an unsigned big-endian whole byte integer.
-        uintle -- an unsigned little-endian whole byte integer.
-        uintne -- an unsigned native-endian whole byte integer.
+        ibe -- a signed big-endian whole byte integer.
+        ile -- a signed little-endian whole byte integer.
+        ine -- a signed native-endian whole byte integer.
+        ube -- an unsigned big-endian whole byte integer.
+        ule -- an unsigned little-endian whole byte integer.
+        une -- an unsigned native-endian whole byte integer.
         filename -- the path of a file which will be opened in binary read-only mode.
 
         Other keyword arguments:
@@ -687,7 +687,7 @@ class Bits:
         if length is None and hasattr(self, 'len') and len(self) != 0:
             length = len(self)
         if length is None or length == 0:
-            raise bitstring.CreationError("A non-zero length must be specified with a uintbe initialiser.")
+            raise bitstring.CreationError("A non-zero length must be specified with a ube initialiser.")
         self._bitstore = helpers.int2bitstore(uintbe, length, False)
 
     def _getuintbe(self) -> int:
@@ -704,7 +704,7 @@ class Bits:
         if length is None and hasattr(self, 'len') and len(self) != 0:
             length = len(self)
         if length is None or length == 0:
-            raise bitstring.CreationError("A non-zero length must be specified with a intbe initialiser.")
+            raise bitstring.CreationError("A non-zero length must be specified with an ibe initialiser.")
         self._bitstore = helpers.int2bitstore(intbe, length, True)
 
     def _getintbe(self) -> int:
@@ -720,7 +720,7 @@ class Bits:
         if length is None and hasattr(self, 'len') and len(self) != 0:
             length = len(self)
         if length is None or length == 0:
-            raise bitstring.CreationError("A non-zero length must be specified with a uintle initialiser.")
+            raise bitstring.CreationError("A non-zero length must be specified with a ule initialiser.")
         self._bitstore = helpers.intle2bitstore(uintle, length, False)
 
     def _getuintle(self) -> int:
@@ -736,7 +736,7 @@ class Bits:
         if length is None and hasattr(self, 'len') and len(self) != 0:
             length = len(self)
         if length is None or length == 0:
-            raise bitstring.CreationError("A non-zero length must be specified with an intle initialiser.")
+            raise bitstring.CreationError("A non-zero length must be specified with an ile initialiser.")
         self._bitstore = helpers.intle2bitstore(intle, length, True)
 
     def _getintle(self) -> int: